Move function, so WaveTrack does not depend on TracksBehaviorsPrefs
