+++ src/prefs/TracksBehaviorsPrefs.cpp
@@ -135,18 +135,3 @@ PrefsPanel::Registration sAttachment{ "TracksBehaviors",
    { "Tracks" }
 };
 }
-
-// Bug 825 is essentially that SyncLock requires EditClipsCanMove.
-// SyncLock needs rethinking, but meanwhile this function
-// fixes the issues of Bug 825 by allowing clips to move when in
-// SyncLock.
-bool GetEditClipsCanMove()
-{
-   bool mIsSyncLocked;
-   gPrefs->Read(wxT("/GUI/SyncLockTracks"), &mIsSyncLocked, false);
-   if( mIsSyncLocked )
-      return true;
-   bool editClipsCanMove;
-   gPrefs->Read(wxT("/GUI/EditClipCanMove"), &editClipsCanMove, false);
-   return editClipsCanMove;
-}
