+++ src/prefs/TracksBehaviorsPrefs.h
@@ -39,6 +39,4 @@ class AUDACITY_DLL_API TracksBehaviorsPrefs final : public PrefsPanel
 
 extern AUDACITY_DLL_API ChoiceSetting TracksBehaviorsSolo;
 
-AUDACITY_DLL_API bool GetEditClipsCanMove();
-
 #endif
